wrist now works

+++ Logiciel/STM/STM/Core/Src/UART_Com.c
@@ -28,7 +28,7 @@ void UART_Send(uint8_t M_0, uint8_t M_1, uint8_t M_2, uint8_t M_3, uint8_t M_4){
     UART_Com(M_3);
     UART_Com(M_4);
 
-    CheckSum = ('G' + 'O' + M_0 + M_1 + M_2 + M_3 + M_4);
+    CheckSum = (uint8_t)('G' + 'O' + M_0 + M_1 + M_2 + M_3 + M_4);
     UART_Com(CheckSum);
 }
